Create README.md
So, this project is some sort of attempt to (-make-world-better-place-) reform existing system of sells in all the world. Here only android partof all system, but also exists(or maybe not yet) server part. It wouldn't be useful in sence of code, it wouldn't be useful in any sence actually. It's only one more startup. So, (-get-out-of-here-stalker-) wellcome :^D